requirements.txt
sentence-transformers
faiss-cpu
ollama
pymupdf
python-docx